pyproject.toml
[project]
name = "llm-redteaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.15",
    "langgraph>=1.1.9",
    "langgraph-cli[inmem]>=0.4.23",
    "pypdf>=6.10.2",
    "pyrit>=0.13.0",
    "python-dotenv>=1.2.2",
]
